音声まで含めた動画が生成できる状態になった

+++ app/src/main/java/tokyo/meson/moviecapturetest/Data.kt
@@ -1,4 +1,4 @@
 package tokyo.meson.moviecapturetest
 
 data class FrameData(val data: ByteArray, val timestamp: Long)
-data class AudioData(val data: ShortArray, val timestamp: Long)
+data class AudioData(val data: ByteArray, val timestamp: Long)
